Unable to renew expired passwords
When password is renewed on expiry, pam_unix
module updates /etc/shadow with the new password,
bypassing configd, and therefore, encrypted passwords
in /etc/shadow and configd are not in sync.

Add a custom vyatta pam module that renews the
password via configd and bypasses pam_unix. The
custom pam module executes a python script that
runs the CLI commands for setting the
user's new password.

VRVDR-54100