Bug fixes in opasswd file
Fixed the counts in opasswd file which were
incorrectly being updated when password history
config was being changed. The checks if password
history is configured where also incorrect and
are fixed. Fixed the issue where the current
encrypted password for all users is not written
to the opasswd file when the file is created
for the first time.

VRVDR-54100