Password history config changes default max password age
With password reuse feature, when password
expiration is not configured, we are setting
the maximum password age to -1, which disables
the check for password expiry, and overrides
the default value of 99999. The value in the
YANG file is a range between 1-99999, and
therefore, the fix sets the default value to
99999 when password expiration is not to
configured.

VRVDR-55028