Remove special group added for password change
Use RPC implementation for password change that
provides metadata on user calling the rpc call.
Define a change-password rpc call with
current-password and new-password as it's input.
The RPC script obtains the user from the
VCI_RPC_METADATA environment variable, which
would be the user making the rpc call.
This approach allows all level users, superuser,
admin and operators to change their password, without
the need for a special group for updating passwords.

The PAM module is also updated to prompt the user
for current password. To verify the current password
the PAM module requires another RPC call, verify-password,
with current-password as the input. After verifying
the current password, the PAM module prompts for
'New password', and uses the change-password RPC
call for updating the password.

VRVDR-55918